passwordgenerator/requirements.txt
copyclipboard
